backend/requirements.txt
celery==5.2.7
flower==1.2.0
Django==4.2
django-colorfield==0.11.0
django-cors-headers==3.13.0
django-filter==23.3
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
drf-extra-fields==3.7.0
flake8==7.0.0
gunicorn==20.1.0
isort==5.13.2
Pillow==9.0.0
psycopg2-binary==2.9.3
PyJWT==2.8.0
python-dotenv==1.0.0
pyyaml==6.0
redis==4.4.0
